refactor(F258): resolve BLE implementation review
Why:
- Keep the CoreBluetooth state machine intact while separating delegate callbacks and support types into reviewable files below the size limit.
- Make Phase B adapter composition and subscription cleanup explicit acceptance criteria instead of leaving review observations as informal debt.

[Sol/gpt-5.6-sol🐾]
