+++ docs/ROADMAP.md
@@ -91,4 +91,4 @@ created: 2026-02-26
 | F254 | Side-Effect Freshness Gate — 副作用出口 freshness 拦截（猫发消息前自动检查"有没有你没看过的新消息"→有就 hold 让猫先看；三 surface 一子系统：A Freshness Gate held draft + B Content-Free Notice 防无视 + C Runtime Descriptor 结构化；基于 DeliveryCursorStore seq 游标 + F233 事件流；来源 Raft 0.63.7 teardown + opus-48 修正） | in-progress | Ragdoll (Opus-4.6) | internal (operator 2026-06-27 signoff；Raft teardown 提炼) | [F254](features/F254-side-effect-freshness-gate.md) |
 | F255 | Auto Dream — 会做梦的猫/猫猫日记（后台做梦引擎给 F231 闲置养熟循环"通水" + 前台挂 F229 日记本/provoke 气泡；F255 produce / F229 surface；小而完整垂直切片不做脚手架）| spec | Ragdoll (opus-48) | internal (operator 2026-06-29 "现在立项！") | [F255](features/F255-auto-dream.md) |
 | F256 | Memory Search Strategy Evolution — 从被动召回到主动探索（session hook 升级 + skill link + retrieval expansion hints 投影到默认搜索 + F242 extractor 扩展 + eval 闭环；operator prompting 策略沉淀为猫自主搜索策略）| spec | Ragdoll (opus-4.6) | internal (operator 2026-06-29 confirmed direction) | [F256](features/F256-memory-search-strategy-evolution.md) |
-| F258 | BLE Physical Event Limb — 可审计的物理事件总线（复用 F126 控制面；BLE Central / GATT、显式绑定、标准传感器、按钮事件、平台原生 helper；proximity 不作为敏感操作认证） | in-progress (Phase A implementation complete; AC-A8 hardware acceptance + implementation review pending) | Maine Coon Sol (GPT-5.6 Sol) | internal (operator 2026-07-14「立项开搞」) | [F258](features/F258-ble-physical-event-limb.md) |
+| F258 | BLE Physical Event Limb — 可审计的物理事件总线（复用 F126 控制面；BLE Central / GATT、显式绑定、标准传感器、按钮事件、平台原生 helper；proximity 不作为敏感操作认证） | in-progress (Phase A implementation reviewed; AC-A8 hardware acceptance pending) | Maine Coon Sol (GPT-5.6 Sol) | internal (operator 2026-07-14「立项开搞」) | [F258](features/F258-ble-physical-event-limb.md) |
